Authentication Module

+++ src/main/java/com/logan/service/CollectionService.java
@@ -1,11 +1,8 @@
 package com.logan.service;
 
-import java.text.DateFormat;
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
 import javax.transaction.Transactional;
 
@@ -14,9 +11,7 @@ import org.springframework.stereotype.Service;
 
 import com.logan.dao.CollectionsDAO;
 import com.logan.domain.Collection;
-import com.logan.domain.Users;
 import com.logan.repository.CollectionRepository;
-import com.logan.repository.UsersRepository;
 
 @Service
 @Transactional
@@ -33,7 +28,7 @@ public class CollectionService {
 		return list;
 	}
 	
-	public List<Collection> getCollectionList(Integer userId, Date fromDate, Date toDate) throws ParseException{
+	public List<Collection> getCollectionList(Long userId, Date fromDate, Date toDate) throws ParseException{
 		List<Collection> list = collectionRepository.getCollectionByUser(userId, fromDate, toDate);
 		return list;
 	}
